+++ src/dao/SearchDAO.java
@@ -34,7 +34,7 @@ public class SearchDAO {
 				ChuyenDetails chuyen = new ChuyenDetails();
 //				available++;
 //				System.out.println("Khanh");
-				chuyen.setID(rs.getFloat("routeID"));
+				chuyen.setID(rs.getInt("routeID"));
 //				System.out.println(rs.getFloat("ID"));
 				chuyen.setSTT(rs.getInt("SERIAL"));
 //				System.out.println(rs.getInt("STT"));
